+++ config/routes.rb
@@ -6,8 +6,8 @@ Scrapbook::Engine.routes.draw do
   #   resources :pages, id: /.+/, only: %i[new create edit update]
   # end
 
-  root 'pages#index'
+  root 'pages#show'
 
-  get '.raw/pages/*id', to: 'pages#raw', constraints: {id: /.*/}, as: :raw_page
+  get '.raw/pages(/*id)', to: 'pages#raw', constraints: {id: /.*/}, as: :raw_page
   get '*id', to: 'pages#show', constraints: {id: /.*/}, as: :short_page
 end