+++ spec/controllers/scrapbook/pages_controller_spec.rb
@@ -5,82 +5,89 @@ require 'rails_helper'
 RSpec.describe Scrapbook::PagesController, :aggregate_failures do
   routes { Scrapbook::Engine.routes }
 
-  describe '#index' do
+  describe '#show' do
     let(:scrapbook) { Scrapbook::Scrapbook.new(PathnameHelpers.new.scrapbook_root) }
 
-    it 'gets the listing of the pages directory for the specified book' do
-      get :index, params: {'.book': 'scrapbook'}
-      pathname = scrapbook.pages_pathname
+    it 'renders the show template for the root path (no "id" parameter)' do
+      get :show, params: {'.book': 'scrapbook'}
 
-      expect(pathname).not_to be_empty
-      expect(response).to have_http_status(:ok).and render_template('pages')
-      expect(template_locals).to include(scrapbook: scrapbook, pathname: pathname)
+      expect(response).to have_http_status(:ok).and render_template(:show)
+      expect(template_locals).to include(
+        scrapbook: scrapbook,
+        pathname: scrapbook.pages_pathname
+      )
     end
 
-    it 'gets the listing of the specified sub directory for the specified book' do
-      get :index, params: {'.book': 'scrapbook', path: 'components'}
-      pathname = scrapbook.pages_pathname.join('components')
+    it 'renders the show template for a folder with a template file' do
+      get :show, params: {'.book': 'scrapbook', id: 'components/folder_name'}
 
-      expect(pathname).not_to be_empty
-      expect(response).to have_http_status(:ok).and render_template(:index)
-      expect(template_locals).to include(scrapbook: scrapbook, pathname: pathname)
+      expect(response).to have_http_status(:ok).
+        and render_template('layouts/scrapbook/application').
+        and render_template(:show)
+      expect(template_locals).to include(
+        scrapbook: scrapbook,
+        pathname: scrapbook.pages_pathname.join('components/folder_name')
+      )
     end
 
-    it "returns a 404 error when the book doesn't exist" do
-      get :index, params: {'.book': 'missing'}
-      expect(response).to have_http_status(:not_found)
-    end
-
-    it "returns a 404 error when the path doesn't exist" do
-      get :index, params: {'.book': 'scrapbook', path: 'missing'}
-      expect(response).to have_http_status(:not_found)
-    end
-
-    context 'when no book is specified' do
-      it "returns a 404 error when the path doesn't exist" do
-        get :index, params: {path: 'my/cool/path'}
-        expect(response).to have_http_status(:not_found)
-      end
-    end
-  end
-
-  describe '#show' do
-    it 'renders the show template when the specified template exists in the specified scrapbook' do
-      get :show, params: {'.book': 'scrapbook', id: 'welcome'}
-      expect(response).to render_template('show')
-      expect(response).to render_template('layouts/scrapbook/application')
-    end
-
-    it 'renders the show template when a template exists in the specified scrapbook that matches a request with an "html" extension' do # rubocop:disable Layout/LineLength
-      get :show, params: {'.book': 'scrapbook', id: 'welcome.html'}
-      expect(response).to render_template('show')
-      expect(response).to render_template('layouts/scrapbook/application')
-    end
-
-    it 'renders a directory listing for the specified folder in the specified scrapbook' do
+    it 'renders the show template for a folder without a template file' do
       get :show, params: {'.book': 'scrapbook', id: 'components/folder_name/sub_stuff'}
-      expect(response).to render_template('scrapbook/pages/index')
-      expect(response).to render_template('layouts/scrapbook/application')
+
+      expect(response).to have_http_status(:ok).
+        and render_template('layouts/scrapbook/application').
+        and render_template(:show)
+      expect(template_locals).to include(
+        scrapbook: scrapbook,
+        pathname: scrapbook.pages_pathname.join('components/folder_name/sub_stuff')
+      )
+    end
+
+    it 'renders the show template for the specified template' do
+      get :show, params: {'.book': 'scrapbook', id: 'welcome'}
+
+      expect(response).to have_http_status(:ok).
+        and render_template('layouts/scrapbook/application').
+        and render_template(:show)
+      expect(template_locals).to include(
+        scrapbook: scrapbook,
+        pathname: scrapbook.pages_pathname.join('welcome')
+      )
+    end
+
+    it 'renders the show template when a template exists that matches a request with an "html" extension' do
+      get :show, params: {'.book': 'scrapbook', id: 'welcome.html'}
+
+      expect(response).to have_http_status(:ok).
+        and render_template('layouts/scrapbook/application').
+        and render_template(:show)
+      expect(template_locals).to include(
+        scrapbook: scrapbook,
+        pathname: scrapbook.pages_pathname.join('welcome.html')
+      )
     end
 
     it 'renders the show template when a non-template file is requested' do
       get :show, params: {'.book': 'scrapbook', id: 'assets/fireworks.jpg'}
-      expect(response).to have_http_status(:ok)
-      expect(response).to render_template('show')
-      expect(response).to render_template('layouts/scrapbook/application')
-    end
 
-    it 'prefers to render the show template over a directory or file' do
-      get :show, params: {'.book': 'scrapbook', id: 'components/folder_name'}
-      expect(response).to render_template('show')
-      expect(response).to render_template('layouts/scrapbook/application')
+      expect(response).to have_http_status(:ok).
+        and render_template('layouts/scrapbook/application').
+        and render_template(:show)
+      expect(template_locals).to include(
+        scrapbook: scrapbook,
+        pathname: scrapbook.pages_pathname.join('assets/fireworks.jpg')
+      )
     end
 
     it "renders the show template when the path doesn't exist" do
       get :show, params: {'.book': 'scrapbook', id: 'missing'}
-      expect(response).to have_http_status(:ok)
-      expect(response).to render_template('show')
-      expect(response).to render_template('layouts/scrapbook/application')
+
+      expect(response).to have_http_status(:ok).
+        and render_template('layouts/scrapbook/application').
+        and render_template(:show)
+      expect(template_locals).to include(
+        scrapbook: scrapbook,
+        pathname: scrapbook.pages_pathname.join('missing')
+      )
     end
 
     it "returns a 404 error when the scrapbook doesn't exist" do
@@ -121,9 +128,29 @@ RSpec.describe Scrapbook::PagesController, :aggregate_failures do
       expect(response).to render_template('layouts/scrapbook/host_application')
     end
 
-    it "returns a 404 error when the path doesn't exist" do
-      get :raw, params: {'.book': 'scrapbook', id: 'missing'}
-      expect(response).to have_http_status(:not_found)
+    context "with view rendering" do
+      render_views
+
+      it 'renders the message about creating templates for selected folder that does not have one' do
+        get :raw, params: {'.book': 'scrapbook', id: 'components'}
+
+        expect(response).to render_template('pages')
+        expect(response).to render_template('layouts/scrapbook/host_application')
+        expect(response).not_to include('You can add a template named "pages" to the Scrapbook root directory')
+      end
+
+      it 'renders the message about the root template for scrapbooks missing that file' do
+        get :raw, params: {'.book': 'scrapbook'}
+
+        expect(response).to render_template('pages')
+        expect(response).to render_template('layouts/scrapbook/host_application')
+        expect(response.body).to include('You can add a template named "pages" to the Scrapbook root directory')
+      end
+
+      it "returns a 404 error when the path doesn't exist" do
+        get :raw, params: {'.book': 'scrapbook', id: 'missing'}
+        expect(response).to have_http_status(:not_found)
+      end
     end
 
     it "returns a 404 error when the scrapbook doesn't exist" do