+++ spec/routing/scrapbook_routes_spec.rb
@@ -8,7 +8,7 @@ RSpec.describe Scrapbook do
   describe 'the root route (/)' do
     subject { {get: '/'} }
 
-    it { is_expected.to route_to(controller: 'scrapbook/pages', action: 'index') }
+    it { is_expected.to route_to(controller: 'scrapbook/pages', action: 'show') }
   end
 
   describe 'Direct routes to pages' do
@@ -30,11 +30,26 @@ RSpec.describe Scrapbook do
   end
 
   describe 'Raw routes to pages' do
-    it 'correctly routes to the show page path' do
+    it 'correctly routes files to the raw action' do
       expect(get: '/.raw/pages/area/subject/item.xml').to route_to(
         controller: 'scrapbook/pages',
         action: 'raw',
         id: 'area/subject/item.xml'
+      )
+    end
+
+    it 'correctly routes paths to the the raw action' do
+      expect(get: '/.raw/pages/area/subject/item').to route_to(
+        controller: 'scrapbook/pages',
+        action: 'raw',
+        id: 'area/subject/item'
+      )
+    end
+
+    it 'correctly routes the to the raw action when no path is given' do
+      expect(get: '/.raw/pages').to route_to(
+        controller: 'scrapbook/pages',
+        action: 'raw'
       )
     end
   end