+++ CHANGELOG.md
@@ -20,6 +20,12 @@ may consider to be a bug might be behavior a consumer relies upon in their proje
 
 - The installation generators have been updated for the new configuration architecture.
 
+### Fixed
+
+- Rendering the default directory messages for directories that don't have a corresponding
+  template now renders through the iframe to the raw routes. This means they can be styled
+  like the rest of the templates when we add support for customizing the layout.
+
 [Unreleased commits](https://github.com/bfad/scrapbook/compare/v0.2.1...HEAD)
 
 ## 0.2.1 (2022-08-14)